Fix new -Werror compilation errors
Fix the remaining compilation errors when compiling with -Werror:

* Remove `strictfp` keywords: as of release 17, all floating-point
  expressions are evaluated strictly and 'strictfp' is not required
* java.security.AccessController in java.security has been deprecated and marked for removal
* non-varargs call of varargs method with inexact argument type for last parameter
* deprecated item is not annotated with @Deprecated
* deprecated item is not annotated with @Deprecated
* unreachable catch clause

+++ modules/library/metadata/src/main/java/org/geotools/measure/BaseUnitFormatter.java
@@ -445,7 +445,7 @@ public class BaseUnitFormatter implements UnitFormatter {
                     if (token == Token.INTEGER) {
                         long n = readLong(csq, pos);
                         if (n != 1) {
-                            result = result.multiply(n);
+                            result = result.multiply((double) n);
                         }
                     } else if (token == Token.FLOAT) {
                         double d = readDouble(csq, pos);
@@ -462,7 +462,7 @@ public class BaseUnitFormatter implements UnitFormatter {
                     if (token == Token.INTEGER) {
                         long n = readLong(csq, pos);
                         if (n != 1) {
-                            result = result.divide(n);
+                            result = result.divide((double) n);
                         }
                     } else if (token == Token.FLOAT) {
                         double d = readDouble(csq, pos);
@@ -479,7 +479,7 @@ public class BaseUnitFormatter implements UnitFormatter {
                     if (token == Token.INTEGER) {
                         long n = readLong(csq, pos);
                         if (n != 1) {
-                            result = result.shift(n);
+                            result = result.shift((double) n);
                         }
                     } else if (token == Token.FLOAT) {
                         double d = readDouble(csq, pos);
